+++ src/Controller/DefaultController.php
@@ -2,14 +2,9 @@
 
 namespace App\Controller;
 
-use App\Form\Filter;
-use App\Form\FilterFormType;
-use App\Repository\BreedRepository;
-use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use App\Repository\OfferRepository;
 use App\Repository\BreederRepository;
-use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
@@ -27,27 +22,4 @@ class DefaultController extends AbstractController
         ]);
     }
 
-    // @TODO Move to OfferController //
-    #[Route('/nos-annonces', name: 'offers_list')]
-    public function listOffers(
-        Request $request,
-        Filter $filter, 
-        OfferRepository $offerRepository, 
-        PaginatorInterface $paginator): Response
-    {
-        $filter = new Filter();
-        $form = $this->createForm(FilterFormType::class, $filter,['method' => 'GET',]);
-        $form->handleRequest($request);
-
-        $offers = $offerRepository->findAllOffers($filter);
-        $offers = $paginator->paginate(
-            $offers,
-            $request->query->getInt('page', 1),
-            2
-        );
-        return $this->render('offer/offers_list.html.twig', [
-            'filterForm' => $form->createView(),
-            'offers' => $offers,
-        ]);
-    }
 }
